Fix 2 infinite loops if IMC doesn't respond
ACPI code:
The ACPI code is not currently being compiled in by default, but
assuming that it will be at some point, I'm fixing the loop that
waits for the IMC to respond after sending it a command.  The
loop now exits after 500ms, similar to the function in agesa.

Agesa Code:
a 16 bit variable will always be less than 100000.  Change to
be a 32 bit variable.

+++ src/vendorcode/amd/agesa/f15tn/Proc/Fch/Imc/ImcLib.c
@@ -126,7 +126,7 @@ WaitForEcLDN9MailboxCmdAck (
   )
 {
   UINT8    Msgdata;
-  UINT16   Delaytime;
+  UINT32   Delaytime;
 
   Msgdata = 0;
